init: sorobix serber
