+++ requirements.txt
@@ -1,22 +1,37 @@
 aiohttp==3.8.3
 aiohttp-sse-client==0.2.1
 aiosignal==1.3.1
+anyio==3.6.2
 async-timeout==4.0.2
 attrs==22.2.0
-certifi @ file:///croot/certifi_1669749558946/work/certifi
+certifi==2022.12.7
 cffi==1.15.1
 charset-normalizer==2.1.1
+click==8.1.3
+fastapi==0.89.1
 frozenlist==1.3.3
+h11==0.14.0
+httptools==0.5.0
 idna==3.4
 mnemonic==0.20
 multidict==6.0.4
 pycparser==2.21
+pydantic==1.10.4
 PyNaCl==1.5.0
+python-dotenv==0.21.1
+PyYAML==6.0
 requests==2.28.2
 six==1.16.0
+sniffio==1.3.0
+starlette==0.22.0
 stellar-base-sseclient==0.0.21
 stellar-sdk==8.1.1
 toml==0.10.2
 typeguard==2.13.3
+typing_extensions==4.4.0
 urllib3==1.26.14
+uvicorn==0.20.0
+uvloop==0.17.0
+watchfiles==0.18.1
+websockets==10.4
 yarl==1.8.2
